docker_image/requirements.txt
boto3==1.35.74
python-dotenv==1.0.1
fastapi==0.111.1
mangum==0.17.0
uvicorn==0.30.1
pypdf==4.3.1
chromadb==0.5.23
langchain-chroma==0.2.2
langchain-aws~=0.2.11
langchain~=0.3.13
langchain-text-splitters~=0.3.4
langchain-community==0.3.16
